Inline SceneView code into scene window file
