+++ CMakeLists.txt
@@ -64,7 +64,6 @@ set(EDITOR_SRC
     src/editor/editor.cpp
     src/editor/ui/log_window.cpp
     src/editor/ui/main_menu_bar.cpp
-    src/editor/ui/scene_view.cpp
     src/editor/ui/scene_window.cpp
 )
